requirements.txt
simpy==3.0.5
--allow-external PyPubSub
PyPubSub==3.3.0
pytest==2.6.4
pytest-cov==1.8.1
lazy=1.2
